+++ Firestore/Source/Core/FSTQuery.h
@@ -174,6 +174,7 @@ typedef NS_ENUM(NSInteger, FSTRelationFilterOperator) {
  * Initializes a query with all of its components directly.
  */
 - (instancetype)initWithPath:(firebase::firestore::model::ResourcePath)path
+             collectionGroup:(nullable NSString *)collectionGroup
                     filterBy:(NSArray<FSTFilter *> *)filters
                      orderBy:(NSArray<FSTSortOrder *> *)sortOrders
                        limit:(NSInteger)limit
@@ -187,6 +188,16 @@ typedef NS_ENUM(NSInteger, FSTRelationFilterOperator) {
  * @return A new instance of FSTQuery.
  */
 + (instancetype)queryWithPath:(firebase::firestore::model::ResourcePath)path;
+
+/**
+ * Creates and returns a new FSTQuery.
+ *
+ * @param path The path to the collection to be queried over.
+ * @param collectionGroup The collection group to be queried over.
+ * @return A new instance of FSTQuery.
+ */
++ (instancetype)queryWithPath:(firebase::firestore::model::ResourcePath)path
+              collectionGroup:(nullable NSString *)collectionGroup;
 
 /**
  * Returns the list of ordering constraints that were explicitly requested on the query by the
@@ -244,8 +255,18 @@ typedef NS_ENUM(NSInteger, FSTRelationFilterOperator) {
  */
 - (instancetype)queryByAddingEndAt:(FSTBound *)bound;
 
+/**
+ * Helper to convert a Collection Group query into a collection query at a specific path. This is
+ * used when executing Collection Group queries, since we have to split the query into a set of
+ * Collection queries at multiple paths.
+ */
+- (instancetype)collectionQueryAtPath:(firebase::firestore::model::ResourcePath)path;
+
 /** Returns YES if the receiver is query for a specific document. */
 - (BOOL)isDocumentQuery;
+
+/** Returns YES if the receiver is a collection group query. */
+- (BOOL)isCollectionGroupQuery;
 
 /** Returns YES if the @a document matches the constraints of the receiver. */
 - (BOOL)matchesDocument:(FSTDocument *)document;
@@ -265,6 +286,9 @@ typedef NS_ENUM(NSInteger, FSTRelationFilterOperator) {
 
 /** The base path of the query. */
 - (const firebase::firestore::model::ResourcePath &)path;
+
+/** The collection group of the query. */
+@property(nonatomic, nullable, strong, readonly) NSString *collectionGroup;
 
 /** The filters on the documents returned by the query. */
 @property(nonatomic, strong, readonly) NSArray<FSTFilter *> *filters;