+++ Firestore/Source/Local/FSTLocalStore.mm
@@ -110,8 +110,10 @@ static const int64_t kResumeTokenMaxAgeSeconds = 5 * 60;  // 5 minutes
     _mutationQueue = [persistence mutationQueueForUser:initialUser];
     _remoteDocumentCache = [persistence remoteDocumentCache];
     _queryCache = [persistence queryCache];
-    _localDocuments = [FSTLocalDocumentsView viewWithRemoteDocumentCache:_remoteDocumentCache
-                                                           mutationQueue:_mutationQueue];
+    _localDocuments =
+        [FSTLocalDocumentsView viewWithRemoteDocumentCache:_remoteDocumentCache
+                                             mutationQueue:_mutationQueue
+                                              indexManager:[persistence indexManager]];
     [_persistence.referenceDelegate addInMemoryPins:&_localViewReferences];
 
     _targetIDGenerator = TargetIdGenerator::QueryCacheTargetIdGenerator(0);
@@ -145,8 +147,10 @@ static const int64_t kResumeTokenMaxAgeSeconds = 5 * 60;  // 5 minutes
     std::vector<FSTMutationBatch *> newBatches = _mutationQueue->AllMutationBatches();
 
     // Recreate our LocalDocumentsView using the new MutationQueue.
-    self.localDocuments = [FSTLocalDocumentsView viewWithRemoteDocumentCache:_remoteDocumentCache
-                                                               mutationQueue:_mutationQueue];
+    self.localDocuments =
+        [FSTLocalDocumentsView viewWithRemoteDocumentCache:_remoteDocumentCache
+                                             mutationQueue:_mutationQueue
+                                              indexManager:[_persistence indexManager]];
 
     // Union the old/new changed keys.
     DocumentKeySet changedKeys;