+++ Firestore/Source/Remote/FSTSerializerBeta.mm
@@ -764,10 +764,16 @@ NS_ASSUME_NONNULL_BEGIN
 - (GCFSTarget_QueryTarget *)encodedQueryTarget:(FSTQuery *)query {
   // Dissect the path into parent, collectionId, and optional key filter.
   GCFSTarget_QueryTarget *queryTarget = [GCFSTarget_QueryTarget message];
-  if (query.path.size() == 0) {
-    queryTarget.parent = [self encodedQueryPath:query.path];
+  const ResourcePath &path = query.path;
+  if (query.collectionGroup) {
+    HARD_ASSERT(path.size() % 2 == 0,
+                "Collection Group queries should be within a document path or root.");
+    queryTarget.parent = [self encodedQueryPath:path];
+    GCFSStructuredQuery_CollectionSelector *from = [GCFSStructuredQuery_CollectionSelector message];
+    from.collectionId = query.collectionGroup;
+    from.allDescendants = YES;
+    [queryTarget.structuredQuery.fromArray addObject:from];
   } else {
-    const ResourcePath &path = query.path;
     HARD_ASSERT(path.size() % 2 != 0, "Document queries with filters are not supported.");
     queryTarget.parent = [self encodedQueryPath:path.PopLast()];
     GCFSStructuredQuery_CollectionSelector *from = [GCFSStructuredQuery_CollectionSelector message];
@@ -805,13 +811,18 @@ NS_ASSUME_NONNULL_BEGIN
   ResourcePath path = [self decodedQueryPath:target.parent];
 
   GCFSStructuredQuery *query = target.structuredQuery;
+  NSString *collectionGroup;
   NSUInteger fromCount = query.fromArray_Count;
   if (fromCount > 0) {
     HARD_ASSERT(fromCount == 1,
                 "StructuredQuery.from with more than one collection is not supported.");
 
     GCFSStructuredQuery_CollectionSelector *from = query.fromArray[0];
-    path = path.Append(util::MakeString(from.collectionId));
+    if (from.allDescendants) {
+      collectionGroup = from.collectionId;
+    } else {
+      path = path.Append(util::MakeString(from.collectionId));
+    }
   }
 
   NSArray<FSTFilter *> *filterBy;
@@ -844,6 +855,7 @@ NS_ASSUME_NONNULL_BEGIN
   }
 
   return [[FSTQuery alloc] initWithPath:path
+                        collectionGroup:collectionGroup
                                filterBy:filterBy
                                 orderBy:orderBy
                                   limit:limit