+++ Firestore/Swift/Tests/API/BasicCompileTests.swift
@@ -90,7 +90,7 @@ func makeRefs(database db: Firestore) -> (CollectionReference, DocumentReference
 }
 
 func makeQuery(collection collectionRef: CollectionReference) -> Query {
-  let query = collectionRef.whereField(FieldPath(["name"]), isEqualTo: "Fred")
+  var query = collectionRef.whereField(FieldPath(["name"]), isEqualTo: "Fred")
     .whereField("age", isGreaterThanOrEqualTo: 24)
     .whereField("tags", arrayContains: "active")
     .whereField(FieldPath(["tags"]), arrayContains: "active")
@@ -98,6 +98,9 @@ func makeQuery(collection collectionRef: CollectionReference) -> Query {
     .order(by: FieldPath(["age"]))
     .order(by: "name", descending: true)
     .limit(to: 10)
+
+  // collectionGroup query.
+  query = collectionRef.db.collectionGroup("collection")
 
   return query
 }